+++ test/schemas/one_of_ref_links_schema.json
@@ -1,0 +1,16 @@
+{ "$schema": "http://json-schema.org/draft-04/schema#"
+, "type": "object"
+, "properties":
+  { "links" :
+    { "type" : "array"
+    , "items" :
+      { "type" : "object"
+      , "oneOf" :
+        [ { "$ref" : "self_link_schema.json"}
+        , { "$ref" : "up_link_schema.json" }
+        ]
+      }
+    }
+  }
+}
+
